feat: 방 생성, 조회 api 구현

+++ backend/BE/src/main/java/com/a404/duckonback/repository/ArtistFollowRepository.java
@@ -14,10 +14,10 @@ import java.util.List;
 @Repository
 public interface ArtistFollowRepository extends JpaRepository<ArtistFollow, ArtistFollowId> {
     List<ArtistFollow> findByUser_Id(Long id);
-    List<ArtistFollow> findByArtist_ArtistId(Integer artistId);
-    boolean existsByUser_IdAndArtist_ArtistId(Long id, Integer artistId);
-    void deleteByUser_IdAndArtist_ArtistId(Long id, Integer artistId);
+    List<ArtistFollow> findByArtist_ArtistId(Long artistId);
+    boolean existsByUser_IdAndArtist_ArtistId(Long id, Long artistId);
+    void deleteByUser_IdAndArtist_ArtistId(Long id, Long artistId);
     
     // 특정 아티스트 팔로워 수 조회
-    long countByArtist_ArtistId(Integer artistId);
+    long countByArtist_ArtistId(Long artistId);
 }
